News_App_Project/requirements.txt
Django==4.2.27
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
mysqlclient==2.1.1
Pillow==11.3.0
requests==2.31.0
